--- src/main/java/mike/gui/ViewInterface.java
@@ -1,5 +1,0 @@
-package mike.gui;
-
-public interface ViewInterface {
-
-}
