+++ src/main/java/mike/gui/editBox.java
@@ -16,17 +16,19 @@ import javax.swing.border.Border;
 import javax.swing.border.CompoundBorder;
 import javax.swing.border.EmptyBorder;
 
+import mike.controller.Controller;
 import mike.datastructures.Entity;
+import mike.view.GUIView;
 
 public class editBox {
 	
 	private JLabel newBox;
 	
-	editBox (JLabel label) {
-		Entity e = Controller.getClasses().copyEntity(label.getName());
+	public editBox (JLabel label) {
+		Entity e = Controller.getModel().copyEntity(label.getName());
 		
 		// Get deletion image
-		URL p = GUI.class.getResource("xmark.jpg");
+		URL p = GUIView.class.getResource("..\\gui\\resources\\xmark.jpg");
 		ImageIcon icon = new ImageIcon(new ImageIcon(p).getImage().getScaledInstance(16, 16, Image.SCALE_DEFAULT));
 		
 		// Create entire editBox
@@ -75,46 +77,3 @@ public class editBox {
 		return newBox;
 	}
 }
-//Entire label		
-/*JLabel newview = new JLabel();
-BoxLayout lay = new BoxLayout(newview, BoxLayout.Y_AXIS);
-newview.setLayout(lay);
-
-// Line 1
-JPanel line = new JPanel();
-line.setLayout(new BoxLayout(line, BoxLayout.X_AXIS));
-line.setBackground(Color.LIGHT_GRAY);
-JButton test = new JButton(icon);
-JLabel className = new JLabel("Class: ");
-JTextField test3 = new JTextField("this is text field");
-test.setAlignmentX(Component.LEFT_ALIGNMENT);
-className.setAlignmentX(Component.LEFT_ALIGNMENT);
-test3.setAlignmentX(Component.LEFT_ALIGNMENT);
-line.add(test);
-line.add(className);
-line.add(test3);
-line.setAlignmentX(Component.LEFT_ALIGNMENT);
-line.setBorder(BorderFactory.createEmptyBorder(0, 0, 5, 0));
-
-// Line 2
-JPanel line2 = new JPanel();
-line2.setLayout(new BoxLayout(line2, BoxLayout.X_AXIS));
-line2.setBackground(Color.LIGHT_GRAY);
-JButton test2 = new JButton("hi");
-line2.add(test2);
-line2.setAlignmentX(Component.LEFT_ALIGNMENT);
-line2.setBorder(BorderFactory.createEmptyBorder(5, 0, 3, 0));
-
-// Line 3
-JPanel line3 = new JPanel();
-line3.setLayout(new BoxLayout(line3, BoxLayout.X_AXIS));
-line3.setBackground(Color.LIGHT_GRAY);
-JButton test4 = new JButton("hi once again");
-JTextField test5 = new JTextField("this is another text field");
-line3.add(test4);
-line3.add(test5);
-line3.setAlignmentX(Component.LEFT_ALIGNMENT);
-line3.setBorder(BorderFactory.createEmptyBorder(3, 0, 0, 0));
-
-//newview.add(new JSeparator(JSeparator.HORIZONTAL));
-*/
