+++ src/main/java/mike/gui/resources/xmark.jpg
