+++ src/main/java/mike/view/CLIView.java
@@ -1,4 +1,4 @@
-package mike.cli;
+package mike.view;
 
 import java.io.File;
 import java.io.IOException;
@@ -6,15 +6,16 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Scanner;
 
-import mike.datastructures.Classes;
+import mike.datastructures.Model;
 import mike.datastructures.Relationship.Type;
-import mike.gui.ViewInterface;
+import mike.HelperMethods;
 
-public class CommandLine extends HelperMethods implements ViewInterface {
-	private static Classes classes = new Classes();
+public class CLIView {
+	private Model classes;
 	
-	public CommandLine() {
+	public CLIView(Model classModel) {
     // Initialize variables
+		this.classes = classModel;
 		boolean prompt = false;
 		String[] commandUsage = {
 			"\n  save <name>.json (optional <path>)", 
@@ -87,7 +88,7 @@ public class CommandLine extends HelperMethods implements ViewInterface {
 							else {
 								path = Paths.get(System.getProperty("user.dir") + "\\" + commands[1]);  
 							}
-							save(path, classes);
+							HelperMethods.save(path, classes);
 							System.out.println("File saved at: " + path.toString());
 							prompt = false;
 						}
@@ -104,7 +105,7 @@ public class CommandLine extends HelperMethods implements ViewInterface {
 							else {
 								path = Paths.get(System.getProperty("user.dir") + "\\" + commands[2] + "\\" + commands[1]);  
 							}
-							save(path, classes);
+							HelperMethods.save(path, classes);
 							System.out.println("File saved at: " + path.toString());
 							prompt = false;
 						}
@@ -134,7 +135,7 @@ public class CommandLine extends HelperMethods implements ViewInterface {
 									path = Paths.get(System.getProperty("user.dir") + "\\" + commands[1]);  
 								}
 								
-								load(path, classes);
+								HelperMethods.load(path, classes);
 								prompt = false;
 							}
 						}
@@ -330,7 +331,7 @@ public class CommandLine extends HelperMethods implements ViewInterface {
 							System.out.println(errorMessage + commandUsage[16] + "\n");
 						} else {
 							System.out.println();
-							listClasses(classes);
+							HelperMethods.listClasses(classes);
 							System.out.println();							
 						}
 					} else if (commands[1].equals("relationships")) {
@@ -338,7 +339,7 @@ public class CommandLine extends HelperMethods implements ViewInterface {
 							System.out.println(errorMessage + commandUsage[17] + "\n");
 						} else {
 							System.out.println();							
-							listRelationships(classes);
+							HelperMethods.listRelationships(classes);
 							System.out.println();							
 						}
 					} else if (commands[1].equals("all")) {
@@ -346,9 +347,9 @@ public class CommandLine extends HelperMethods implements ViewInterface {
 							System.out.println(errorMessage + commandUsage[18] + "\n");
 						} else {
 							System.out.println();							
-							listClasses(classes);
+							HelperMethods.listClasses(classes);
 							System.out.println();
-							listRelationships(classes);
+							HelperMethods.listRelationships(classes);
 							System.out.println();							
 						}
 					} else {
